pyproject.toml
[project]
name = "llm-cost-calculator"
version = "0.1.0"
description = ""
authors = [
    {name = "João Lages",email = "joao.lages@complyadvantage.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit (>=1.50.0,<2.0.0)",
    "gspread (>=6.2.1,<7.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "oauth2client (>=4.1.3,<5.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "odfpy (>=1.4.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
